Move storage utils to root level test/utils
