+++ framework/test/jest/integration/utils/storage.js
@@ -17,7 +17,7 @@
 const {
 	StorageSandbox,
 	clearDatabaseTable,
-} = require('../../../mocha/common/storage_sandbox');
+} = require('../../../utils/storage/storage_sandbox');
 
 const getAccount = async (storage, address) =>
 	storage.entities.Account.getOne({ address }, { extended: true });
