+++ framework/test/mocha/integration/blocks/process/process.js
@@ -17,7 +17,7 @@
 const async = require('async');
 const application = require('../../../common/application');
 const modulesLoader = require('../../../common/modules_loader');
-const clearDatabaseTable = require('../../../common/storage_sandbox')
+const clearDatabaseTable = require('../../../../utils/storage/storage_sandbox')
 	.clearDatabaseTable;
 const loadTables = require('./process_tables_data.json');
