+++ framework/test/mocha/unit/components/storage/entities/account.js
@@ -24,8 +24,8 @@ const {
 		NonSupportedOperationError,
 	},
 } = require('../../../../../../src/components/storage');
-const storageSandbox = require('../../../../common/storage_sandbox');
-const seeder = require('../../../../common/storage_seed');
+const storageSandbox = require('../../../../../utils/storage/storage_sandbox');
+const seeder = require('../../../../../utils/storage/storage_seed');
 const accountFixtures = require('../../../../fixtures').accounts;
 
 describe('Account', () => {
