+++ framework/test/mocha/unit/components/storage/entities/block.js
@@ -24,7 +24,7 @@ const {
 	},
 } = require('../../../../../../src/components/storage');
 
-const storageSandbox = require('../../../../common/storage_sandbox');
+const storageSandbox = require('../../../../../utils/storage/storage_sandbox');
 const transactionsFixtures = require('../../../../fixtures').transactions;
 const blocksFixtures = require('../../../../fixtures/blocks');
