+++ framework/test/mocha/unit/components/storage/entities/transaction.js
@@ -23,8 +23,8 @@ const {
 		NonSupportedOptionError,
 	},
 } = require('../../../../../../src/components/storage');
-const storageSandbox = require('../../../../common/storage_sandbox');
-const seeder = require('../../../../common/storage_seed');
+const storageSandbox = require('../../../../../utils/storage/storage_sandbox');
+const seeder = require('../../../../../utils/storage/storage_seed');
 const transactionsFixtures = require('../../../../fixtures').transactions;
 
 const numSeedRecords = 5;
