+++ framework/test/mocha/unit/modules/chain/components/storage/entities/account.js
@@ -22,8 +22,8 @@ const {
 const {
 	Account,
 } = require('../../../../../../../../src/modules/chain/components/storage/entities');
-const storageSandbox = require('../../../../../../common/storage_sandbox');
-const seeder = require('../../../../../../common/storage_seed');
+const storageSandbox = require('../../../../../../../utils/storage/storage_sandbox');
+const seeder = require('../../../../../../../utils/storage/storage_seed');
 const accountFixtures = require('../../../../../../fixtures').accounts;
 
 const defaultCreateValues = {
