+++ framework/test/mocha/unit/modules/chain/components/storage/entities/transaction.js
@@ -21,8 +21,8 @@ const {
 const {
 	Transaction,
 } = require('../../../../../../../../src/modules/chain/components/storage/entities');
-const storageSandbox = require('../../../../../../common/storage_sandbox');
-const seeder = require('../../../../../../common/storage_seed');
+const storageSandbox = require('../../../../../../../utils/storage/storage_sandbox');
+const seeder = require('../../../../../../../utils/storage/storage_seed');
 const transactionsFixtures = require('../../../../../../fixtures').transactions;
 
 const TRANSACTION_TYPES = {
